f3at: before javadoc

+++ domain_model/src/main/java/entities/Transaction.java
@@ -1,16 +1,12 @@
 package entities;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-public record Transaction(UUID transId, UUID accountId, double amount, LocalDateTime transTime) {
-    public Transaction(UUID accountId, double amount) {
+public record Transaction(UUID transId, UUID accountId, BigDecimal amount, LocalDateTime transTime) {
+    public Transaction(UUID accountId, BigDecimal amount) {
         this(UUID.randomUUID(), accountId, amount, LocalDateTime.now());
     }
 }
 
-//public record Transaction(UUID transId, UUID accountId, double amount, LocalDateTime transTime) {
-//    public Transaction(UUID accountId, double amount){
-//        this(UUID.randomUUID(), accountId, amount, LocalDateTime.now());
-//    }
-//}
